+++ Sekoia.io/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2025-07-30 - 2.68.5
+
+### Added
+
+- Add debugging logs in triggers
+
+### Fixed
+
+- Handle empty `version` in `metadata`.
+
 ## 2025-06-03 - 2.68.4
 
 ### Fixed
@@ -19,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improve the description of the case identifier argument in the `Attach Alert to Case` action and `Add events to Case` action.
 - Improve the description of the cases argument in the `Get Alert` action.
+
 
 ## 2025-05-28 - 2.68.2